Safari-611.1.10/LayoutTests/js/dom/Promise-chain-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fulfilled
PASS result is "hello"
PASS fulfilled
PASS result is "hello2"
PASS rejected
PASS result is "error"
PASS rejected
PASS result is "error2"
PASS fulfilled
PASS result is "recovered"
PASS operation is "asynchronous"
PASS successfullyParsed is true

TEST COMPLETE
